References WHO

Reference tables, combined from the WHO Multicentre Growth Reference Study (MGRS) (ages 0-5 years) and the WHO 2007 reference (5-19 years).

Format

A data frame with seven variables:

pop

Study Population (always "who")

sub

Subpopulation (always "N")

sex

Sex (M, F)

x

Decimal age, height (cm) or length(cm)

L

Lambda (skewness) curve

M

Median curve

S

Coefficient of variation

Source

WHO Multicentre Growth Reference Study Group. WHO Child Growth Standards based on length/height, weight and age. Acta Paediatr, Suppl. 2006, 450, 76-85.

de Onis M, Onyango AW, Borghi E, Siyam A, Nishida C, Siekmann J. Development of a WHO growth reference for school-aged children and adolescents Bulletin of the World Health Organization, 2007;85:660-7.

Details

The data were fitted by the LMS model. Parameters are stored as type LMS. Tabulated values are point ages.

The naming conventions are as follows:

who.hgt

Length (cm, 0-2 Yrs) or height (cm, 2-19 years)

who.wgt

Weight (kg) for age (0-10 years)

who.bmi

BMI (kg/m^2) for age (0-19 years)

who.hdc

Head circumference (cm) for age (0-5 years)

who.wfh

Weight (kg) for height (65-120 cm)

who.wfl

Weight (kg) for length (45-110 cm)

See also